+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 
+## 1.6.0 - 2025-02-22
+
+* Added
+  * working scaled, scrollbable Board view
+
 ## 1.5.0 - 2025-02-22
 
 * Added